change task3

+++ src/Itmo/lessons1/Task3.java
@@ -2,8 +2,8 @@ package Itmo.lessons1;
 
 public class Task3 {
     public void count(){
-        int number = 10500;
-        double result = ((double) (number/10)/10);
+        double number = 10500;
+        double result =  (number/10)/10;
         System.out.println(result);
     }
 }
